pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caseflow-decision-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.40.0",
  "duckdb>=1.4.4",
  "fastapi>=0.129.0",
  "langgraph>=0.2.20",
  "pillow>=12.1.1",
  "prometheus-client>=0.20",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.12.5",
  "pytesseract>=0.3.13",
  "python-dotenv>=1.2.1",
  "pyyaml>=6.0.3",
  "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
  "black>=26.1.0",
  "httpx>=0.28.1",
  "mypy>=1.19.1",
  "pre-commit>=4.5.1",
  "pytest>=9.0.2",
  "ruff>=0.15.1",
]
notebook = [
  "ipykernel>=7.2.0",
  "jupyter>=1.1.1",
  "pandas>=2.2.0",
]
ml = [
  "langchain-community>=0.3.0",
  "langchain-core>=0.3.0",
  "matplotlib>=3.9",
  "numpy>=2.0.0",
  "pyarrow>=16",
  "scikit-learn>=1.5.0",
]

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "I"]

[tool.black]
line-length = 88

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["caseflow*"]
